+++ pyproject.toml
@@ -18,4 +18,4 @@ dependencies = [
 browser-harness = "run:main"
 
 [tool.setuptools]
-py-modules = ["run", "helpers", "daemon", "admin", "ipc"]
+py-modules = ["run", "helpers", "daemon", "admin", "_ipc"]
